ci: fix the ci - add cast in endian because gcc can do some weird warnings about the uint16_t comparison in assert_equal - fix invalid array check in capabilities.c

+++ sources/pkgs/test/brutal/base/endian.c
@@ -5,17 +5,17 @@ TEST(endian_load_store_be16)
 {
     be_uint16_t value;
     store_be(0x1a2b, &value);
-    uint32_t result = load_be(value);
-    assert_equal(result, 0x1a2b);
-    assert_equal(value._be_data, 0x2b1a);
+    uint16_t result = load_be(value);
+    assert_equal(result, (uint16_t)0x1a2b);
+    assert_equal(value._be_data, (uint16_t)0x2b1a);
 }
 
 TEST(endian_load_store_le16)
 {
     le_uint16_t value;
     store_le(0x1a2b, &value);
-    uint32_t result = load_le(value);
-    assert_equal(result, 0x1a2b);
+    uint16_t result = load_le(value);
+    assert_equal(result, (uint16_t)0x1a2b);
 }
 
 TEST(endian_load_store_be32)
@@ -23,7 +23,7 @@ TEST(endian_load_store_be32)
     be_uint32_t value;
     store_be(0x1a2b3c4d, &value);
     uint32_t result = load_be(value);
-    assert_equal(result, 0x1a2b3c4d);
+    assert_equal(result, (uint32_t)0x1a2b3c4d);
 }
 
 TEST(endian_load_store_le32)
@@ -31,5 +31,5 @@ TEST(endian_load_store_le32)
     le_uint32_t value;
     store_le(0x1a2b3c4d, &value);
     uint32_t result = load_le(value);
-    assert_equal(result, 0x1a2b3c4d);
+    assert_equal(result, (uint32_t)0x1a2b3c4d);
 }
